+++ netbox/dcim/filters.py
@@ -6,7 +6,7 @@ from django.db.models import Q
 from netaddr import EUI
 from netaddr.core import AddrFormatError
 
-from extras.filters import CustomFieldFilterSet
+from extras.filters import CustomFieldFilterSet, LocalConfigContextFilter
 from tenancy.models import Tenant
 from utilities.constants import COLOR_CHOICES
 from utilities.filters import (
@@ -450,7 +450,7 @@ class PlatformFilter(NameSlugSearchFilterSet):
         fields = ['name', 'slug']
 
 
-class DeviceFilter(CustomFieldFilterSet):
+class DeviceFilter(LocalConfigContextFilter, CustomFieldFilterSet):
     id__in = NumericInFilter(
         field_name='id',
         lookup_expr='in'
@@ -597,10 +597,6 @@ class DeviceFilter(CustomFieldFilterSet):
         method='_pass_through_ports',
         label='Has pass-through ports',
     )
-    local_context_data = django_filters.BooleanFilter(
-        method='_local_context_data',
-        label='Has local config context data',
-    )
     tag = TagFilter()
 
     class Meta:
@@ -654,9 +650,6 @@ class DeviceFilter(CustomFieldFilterSet):
 
     def _interfaces(self, queryset, name, value):
         return queryset.exclude(interfaces__isnull=value)
-
-    def _local_context_data(self, queryset, name, value):
-        return queryset.exclude(local_context_data__isnull=value)
 
     def _pass_through_ports(self, queryset, name, value):
         return queryset.exclude(