+++ netbox/extras/forms.py
@@ -330,6 +330,20 @@ class ConfigContextFilterForm(BootstrapMixin, forms.Form):
 
 
 #
+# Filter form for local config context data
+#
+
+class LocalConfigContextFilterForm(BootstrapMixin, forms.Form):
+    local_context_data = forms.NullBooleanField(
+        required=False,
+        label='Has local config context data',
+        widget=StaticSelect2(
+            choices=BOOLEAN_WITH_BLANK_CHOICES
+        )
+    )
+
+
+#
 # Image attachments
 #